requirements.txt
PyQt6
pynput
pycryptodome